Get rows with serological equivalents of an HLA allele from ETRL HLA table

etrl_lookup() takes in a string or character vector of HLA alleles, and returns their serological equivalents as defined in the ETRL HLA tables.

Usage

etrl_lookup(allele)

Arguments

allele

A string or character vector with (an) HLA allele(s).

Value

A data frame with as many rows as there were elements in the input, and four columns:

The allele that was looked up (or NA if it didn't exist)

Its equivalent at the serological split level (or NA if it doesn't have one)

Its equivalent at the serological broad level (or NA if it doesn't have one)

Whether the allele has the public epitope (Bw4 or Bw6) (or NA if it doesn't)

See also

Examples

allele_vec <- c(
  "B15", "B*15:79N", "B*15:YETY",
  "B*15:01:16", "B*15:02", "B*15:85"
)
etrl_lookup(allele_vec)
#> # A tibble: 6 × 4
#>   Allele  `ET MatchDeterminantSplit` `ET MatchDeterminantBroad` Public
#>   <chr>   <chr>                      <chr>                      <chr> 
#> 1 NA      NA                         NA                         NA    
#> 2 NA      NA                         NA                         NA    
#> 3 B*15:XX NA                         B15                        NA    
#> 4 B*15:01 B62                        B15                        Bw6   
#> 5 B*15:02 B75                        B15                        Bw6   
#> 6 B*15:XX NA                         B15                        NA