v2 to v3 HLA nomenclature conversion table

In April 2010, the HLA nomenclature officially changed from "v2" to "v3", which mandates the use of colons (":") to delimit fields. For example: A*01010101 (v2) is since written as A*01:01:01:01. Note though that not all conversions are this straightforward, there are a number of exceptions.

This conversion table tries to collect all standard v2 –> v3 conversions from different data sources (see below).

Usage

v2_to_v3

Format

v2_to_v3

A data frame with 5,151 rows and 2 columns:

v2

v2 version of the allele

v3

v3 version of the allele

Source

Most pre-2010 alleles: https://github.com/ANHIG/IMGTHLA/blob/Latest/Nomenclature_2009.txt

Additional mappings for deleted alleles: https://github.com/ANHIG/IMGTHLA/blob/Latest/Deleted_alleles.txt

Allele-specific & DPB1-specific multiple-allele codes (see 4. and 5. under "Allele Code Update to Version 3 WHO Nomenclature") N.B. Unfortunately this link no longer works https://bioinformatics.bethematchclinical.org/hla-resources/allele-codes/